Allow window ADFS automatic login for windows
This will make the login and redirect to ADFS transparent for the user. It requires that the ADFS site is added to the internet explorers zone for Intranet sites and that the Windows Autologin. Example of setup can be found here https://confluence.atlassian.com/display/SPCON/Setting+your+Browser+Options+for+Automatic+Login